Correctly detect curly and square breace nesting.
Just counting the nesting level does not detect if a curly is closed by
a square and vice versa:
{
  "id" : "Me",
    "value : [ 1, 2 }
}

Counting the levels searately does not detect
the swapping of a curly with a square:
{
  "id" : "Me",
    "value :{
      "data" : [ 1, 2 }
      ]
}

TODO: Support arbitrary nesting level.
The current implemenation works up to 63 levels of nesting, safe for
human edited workflows, but what of generated ones?