eda: New FlexyField class for to parse JSON values which can be different things
Fix memory leak when calling OnCopy multiple times.

Extend FieldType to cover array and map pointers.

The new values added to FlexyType are not compatible with SQLiteType:
any special treatment for interoperability is to be defined.

Since we only add two values, the FlexyTypeEx class has the same layout as
FieldTypeEx, although it may contain values that are not allowed in the latter
(array, map): if more types were to be added in the future a differnt layout would be required.

Lay out the FlexyField structure.

This struct could be derived from FieldValue, but that would increase the
complexity of the implementation, so a customized copy of FieldValue was preferred.

Two new members of the union allow storage of, and easy access to, pointers to
an Array<FlexyField> or a Map<String, FlexyField>, to suport json schemas where
a value can be a scalar, an array or an object (the latter would be parsed and
stored as a Map object).

Access to union members is implemented via properties to allow automatic
setting of the internal state when assignig a value.
If a different internal state is desired, it can be set manually after the value eg:

    FlexyField f {s=aString};
    f.type.mustFree = false
    FlexyField n {b=null, {nil, false}};

Note that the idiom on the last line is currently the only way to create
manualy a FlexyField of type nil.

Token implementation of ordering function.

Arrays compare as smaller than objects and larger than other types,
  but arrays and maps compare as equal to any value of the sam type,
  until a sensible ordering is devised.

Integer and real comparison are implemented without a function call.

Implement OnFree.

Implement OnCopy: auto deallocateof previous data.

Support for serialization/unserialization.

Move repeated code into external helper functions.

These can be reused with FlexyField too.

Implement OnGetString for all types.

    Arrays are formatted on a single line, via helper method..

    Maps are formatted with one element per line if they have more than three
    elements, on a singler line otherwise, via helper method.

    Plain blobs are formatted as a literal string.

 Implement stringify.

 Return a String-casted copy of blob data,
 use Array::OngetString and Map::OnGetString() for array andmap data.

Implement OnGetDataFromString.

Treat blob data as text.
Use a helper method to parse arrays and maps.

ECONParser only does what is needed here when the json object is in the value
part of a key:value entry, so MapGetHelper is introduced to force that behaviour.

Enable customValuefication for FlexyField.

TODO:
New feature: implement a general way to enable (custom|special)Valuefiation for new
classes without editing JSON.ec

Calling OnFree witin OnCopy can cause problems.

If the structure is not initialized yet, calling OnCopy would try to free
memory at random addresses, causing problems if it succeeds.