Extend FieldType to cover array and map pointers.
The new values added to FlexyType are not compatible with SQLiteType:
any special treatment for interoperability is to be defined.

Since we only add two values, the FlexyTypeEx class has the same layout as
FieldTypeEx, although it may contain values that are not allowed in the latter
(array, map): if more types were to be added in the future a differnt layout would be required.