Lay out the FlexyField structure.
This struct could be derived from FieldValue, but that would increase the
complexity of the implementation, so a customized copy of FieldValue was preferred.

Two new members of the union allow storage of, and easy access to, pointers to
an Array<FlexyField> or a Map<String, FlexyField>, to suport json schemas where
a value can be a scalar, an array or an object (the latter would be parsed and
stored as a Map object).

Access to union members is implemented via properties to allow automatic
setting of the internal state when assignig a value.
If a different internal state is desired, it can be set manually after the value eg:

    FlexyField f {s=aString};
    f.type.mustFree = false
    FlexyField n {b=null, {nil, false}};

Note that the idiom on the last line is currently the only way to create
manualy a FlexyField of type nil.