ecere/sys/JSON.ec::StringIndent: null-termination for 'String target'.
cause: The reallocation of 'target' does not zero thetrailing chars and this was taken
care of in PrintObjectNotationString by calling "f.Putc(0);", however this only
works if StringIndent is called directly by PrintObjectNotationString/WriteONObject,
if it is nested in another function, such as OnGetString, this mechanism fails unpredictably.

Fix: ensure that target is properly null-terminated inside StringIndent.